HMW-115 Added support for remaining layers being pulled in via url params. Also updated add data to use the same code as url params logic.
